// public/static/charting_library/charting_library.d.ts
export declare type Nominal<T, Name extends string> = T & {
  /* eslint-disable-next-line jsdoc/require-jsdoc */ [Symbol.species]: Name;
};
// ...
export interface Bar {
  /** Bar time.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   * `time` for daily, weekly, and monthly bars is expected to be a trading day (not session start day) at 00:00 UTC.
   * The library adjusts time according to `session` from {@link LibrarySymbolInfo}.
   */
  time: number;
  /** Opening price */
  open: number;
  /** High price */
  high: number;
  /** Low price */
  low: number;
  /** Closing price */
  close: number;
  /** Trading Volume */
  volume?: number;
}
// ...
export interface CurrencyItem {
  /** Unique ID */
  id: string;
  /** Currency code. @example `USD` */
  code: string;
  /** URL to an image of the currency. SVG logos are preferable, but please make sure that the provided image fits nicely in the currency select control (for raster images the expected size is 24x24px). */
  logoUrl?: string;
  /** Description for the currency */
  description?: string;
}
// ...
export interface DatafeedConfiguration {
  /**
   * List of exchange descriptors.
   * An empty array leads to an absence of the exchanges filter in the Symbol Search list.
   * Use `value=''` if you wish to include all exchanges.
   */
  exchanges?: Exchange[];
  /**
   * List of [resolutions](https://www.tradingview.com/charting-library-docs/latest/core_concepts/Resolution.md) that the chart should support.
   * Each item of the array is expected to be a string that has a specific [format](https://www.tradingview.com/charting-library-docs/latest/core_concepts/Resolution.md#resolution-format).
   * If you set this property to `undefined` or an empty array, the _Resolution_ drop-down menu displays the list of resolutions available for
   * the current symbol ({@link LibrarySymbolInfo.supported_resolutions}).
   *
   * @example
   * `["1", "15", "240", "D", "6M"]` will give you "1 minute, 15 minutes, 4 hours, 1 day, 6 months" in the _Resolution_ drop-down menu.
   */
  supported_resolutions?: ResolutionString[];
  /**
   * Supported unit groups. Each group can have several unit objects.
   *
   * @example
   * ```javascript
   * {
   *     weight: [
   *         { id: 'kg', name: 'kg', description: 'Kilograms' },
   *         { id: 'lb', name: 'lb', description: 'Pounds'},
   *     ]
   * }
   * ```
   */
  units?: Record<string, Unit[]>;
  /**
   * Supported currencies for currency conversion.
   *
   * When a currency code is supplied as a string then the library will automatically convert it to `{ id: value, code: value }` object.
   *
   * @example `["USD", "EUR", "GBP"]`
   * @example `[{ id: "USD", code: "USD", description: "$" }, { id: "EUR", code: "EUR", description: "€" }]`
   */
  currency_codes?: (string | CurrencyItem)[];
  /** Does the datafeed supports marks on bars (`true`), or not (`false | undefined`). */
  supports_marks?: boolean;
  /** Set this one to `true` if your datafeed provides server time (unix time). It is used to adjust Countdown on the Price scale. */
  supports_time?: boolean;
  /** Does the datafeed supports marks on the timescale (`true`), or not (`false | undefined`). */
  supports_timescale_marks?: boolean;
  /**
   * List of filter descriptors.
   *
   * Setting this property to `undefined` or an empty array leads to the absence of filter types in Symbol Search list. Use `value = ''` if you wish to include all filter types.
   * `value` within the descriptor will be passed as `symbolType` argument to {@link IDatafeedChartApi.searchSymbols}
   */
  symbols_types?: DatafeedSymbolType[];
  /**
   * Set it if you want to group symbols in the [Symbol Search](https://www.tradingview.com/charting-library-docs/latest/ui_elements/Symbol-Search.md).
   * Represents an object where keys are symbol types {@link SymbolType} and values are regular expressions (each regular expression should divide an instrument name into 2 parts: a root and an expiration).
   *
   * Sample:
   * ```
   * {
   *   "futures": `/^(.+)([12]!|[FGHJKMNQUVXZ]\d{1,2})$/`,
   *   "stock": `/^(.+)([12]!|[FGHJKMNQUVXZ]\d{1,2})$/`,
   * }
   * ```
   * It will be applied to the instruments with futures and stock as a type.
   * Refer to [Symbol grouping](https://www.tradingview.com/charting-library-docs/latest/ui_elements/Symbol-Search.md#symbol-grouping) for more information.
   */
  symbols_grouping?: Record<string, string>;
}
// ...
export interface DatafeedSymbolType {
  /** Name of the symbol type */
  name: string;
  /** Value to be passed as the `symbolType` argument to `searchSymbols` */
  value: string;
}
// ...
export interface Exchange {
  /** Value to be passed as the `exchange` argument to `searchSymbols` */
  value: string;
  /** Name of the exchange */
  name: string;
  /** Description of the exchange */
  desc: string;
}
// ...
export interface HistoryMetadata {
  /**
   * Optional value indicating to the library that there is no more data on the server.
   */
  noData?: boolean;
  /**
   * Time of the next bar if there is no data in the requested period.
   * It should be set if the requested period represents a gap in the data. Hence there is available data prior to the requested period.
   */
  nextTime?: number | null;
}
// ...
export interface PeriodParams {
  /**
   * Unix timestamp (leftmost requested bar)
   */
  from: number;
  /**
   * Unix timestamp (rightmost requested bar - not inclusive)
   */
  to: number;
  /**
   * The exact amount of bars to load, should be considered a higher priority than `from` if your datafeed supports it
   */
  countBack: number;
  /**
   * Used to identify if it's the first call of getBars
   */
  firstDataRequest: boolean;
}
// ...
export interface SearchSymbolResultItem {
  /** Short symbol name */
  symbol: string;
  /** Full symbol name  */
  full_name: string;
  /** Description */
  description: string;
  /** Exchange name */
  exchange: string;
  /** Symbol ticker name. Should be an unique id */
  ticker?: string;
  /**
   * Type of symbol
   *
   * 'stock' | 'futures' | 'forex' | 'index'
   */
  type: string;
  /**
   * URL of image/s to be displayed as the logo/s for the symbol. The `show_symbol_logos` featureset needs to be enabled for this to be visible in the UI.
   *
   * - If a single url is returned then that url will solely be used to display the symbol logo.
   * - If two urls are provided then the images will be displayed as two partially overlapping
   * circles with the first url appearing on top. This is typically used for FOREX where you would
   * like to display two country flags as the symbol logo.
   *
   * The image/s should ideally be square in dimension. You can use any image type which
   * the browser supports natively. Simple SVG images are recommended.
   *
   * Examples:
   * - `https://yourserver.com/symbolName.svg`
   * - `/images/myImage.png`
   * - `data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3...`
   * - `data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAASABIAAD/4...`
   */
  logo_urls?: [string] | [string, string];
  /**
   * URL of image to be displayed as the logo for the exchange. The `show_exchange_logos` featureset needs to be enabled for this to be visible in the UI.
   *
   * The image should ideally be square in dimension. You can use any image type which
   * the browser supports natively. Simple SVG images are recommended.
   *
   * Examples:
   * - `https://yourserver.com/exchangeLogo.svg`
   * - `/images/myImage.png`
   * - `data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3...`
   * - `data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAASABIAAD/4...`
   */
  exchange_logo?: string;
}
// ...
export interface Unit {
  /** Unique ID */
  id: string;
  /** Short name */
  name: string;
  /** Description */
  description: string;
}
// ...
export type ErrorCallback = (reason: string) => void;
// ...
export type HistoryCallback = (bars: Bar[], meta?: HistoryMetadata) => void;
// ...
export type OnReadyCallback = (configuration: DatafeedConfiguration) => void;
// ...
export type ResolutionString = Nominal<string, "ResolutionString">;
export type ResolveCallback = (symbolInfo: LibrarySymbolInfo) => void;
/** RSS news feed. */
export type RssNewsFeedItem = RssNewsFeedInfo | RssNewsFeedInfo[];
export type SearchSymbolsCallback = (items: SearchSymbolResultItem[]) => void;
// ...
export type SubscribeBarsCallback = (bar: Bar) => void;